+++ app/Http/Controllers/AuditController.php
@@ -17,37 +17,12 @@
 
 namespace App\Http\Controllers;
 
-use App\Models\Activity;
-use Illuminate\Http\JsonResponse;
 use Illuminate\View\View;
-use Yajra\DataTables\DataTables;
 
 class AuditController extends Controller
 {
-    public function index(): View
+    public function __invoke(): View
     {
         return view('audit.index');
-    }
-
-    public function data(DataTables $dataTable): JsonResponse
-    {
-        $activities = Activity::all();
-
-        return $dataTable->collection($activities)
-            ->addColumn('timestamp', function (Activity $activity) {
-                return $activity->present()->created_at;
-            })
-            ->addColumn('time', function (Activity $activity) {
-                return $activity->present()->activityAge;
-            })
-            ->addColumn('status', function (Activity $activity) {
-                return $activity->present()->statusBadge;
-            })
-            ->addColumn('causer', function (Activity $activity) {
-                return $activity->present()->causerUsername;
-            })
-            ->rawColumns(['status'])
-            ->only(['description', 'status', 'time', 'timestamp', 'causer'])
-            ->toJson();
     }
 }
