+++ tests/Feature/Http/Controllers/UserControllerTest.php
@@ -623,50 +623,4 @@ class UserControllerTest extends TestCase
 
         $this->assertSoftDeleted($user);
     }
-
-    /** @test */
-    public function viewers_should_get_error_when_getting_data_tables_data_with_non_AJAX_requests(): void
-    {
-        $this->user->givePermissionTo(Permissions::ViewUsers);
-
-        $this
-            ->actingAs($this->user)
-            ->get(route('users.data'))
-            ->assertForbidden();
-    }
-
-    /** @test */
-    public function users_should_not_get_data_tables_data(): void
-    {
-        $this
-            ->actingAs($this->user)
-            ->ajaxGet(route('users.data'))
-            ->assertForbidden();
-    }
-
-    /** @test */
-    public function viewers_should_get_data_tables_data(): void
-    {
-        $this->user->givePermissionTo(Permissions::ViewUsers);
-
-        $users = User::factory()
-            ->count(3)
-            ->create([
-                'enabled' => 'true',
-            ]);
-
-        $this
-            ->actingAs($this->user)
-            ->ajaxGet(route('users.data'))
-            ->assertSuccessful()
-            ->assertJsonCount(count($users) + 1, 'data')
-            ->assertJsonStructure([
-                'data' => [
-                    '*' => [
-                        'username',
-                        'email',
-                    ],
-                ],
-            ]);
-    }
 }
